requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.10.0
Flask-Login==0.6.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
gunicorn==22.0.0
APScheduler==3.10.4
torch==2.3.1
scikit-learn==1.4.2
statsmodels==0.14.2
arch==7.0.0
joblib==1.4.2
Flask-WTF==1.2.1
WTForms==3.1.2
Werkzeug==3.0.2
itsdangerous==2.2.0
pytz==2024.1
beautifulsoup4==4.12.3
flask-profiler==1.8.1
email-validator==2.2.0